requirements.txt
# Core Django dependencies
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.7

# Authentication & Security
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0

# File handling
Pillow>=10.0.0
django-storages==1.14.2
boto3==1.34.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.0

# PDF Generation
reportlab==4.0.7

# Excel/CSV export
openpyxl==3.1.2
xlsxwriter==3.1.9

# Monitoring & Logging
sentry-sdk==1.38.0

# Development tools
django-debug-toolbar==4.2.0
factory-boy==3.3.0
faker==20.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Frontend build tools (optional)
django-compressor==4.4
django-sass-processor==1.2.2
libsass==0.22.0

# Utilities
django-model-utils==4.3.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Task queue (for async operations)
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# WSGI server for production
gunicorn==21.2.0
whitenoise==6.6.0
